# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Media FFmpeg Service API
  version: 1.0.0
  license:
    name: UNLICENSED
    url: https://choosealicense.com/no-permission/
  description: >
    API REST para almacenamiento permanente de archivos multimedia y procesamiento
    asíncrono con FFmpeg, MongoDB/GridFS y BullMQ. Todo endpoint que usa FFmpeg
    crea un Job. Todo output generado por FFmpeg se registra como Artifact.
servers:
  - url: /api/v1
    description: Base path del servicio
security: []
...
paths:
...
  /video/{mediaId}/extract-frames:
    post:
      tags: [Video]
      summary: Crea un job para extraer frames.
      operationId: extractFrames
      parameters:
        - $ref: '#/components/parameters/MediaId'
      requestBody:
        $ref: '#/components/requestBodies/VideoExtractFrames'
      responses:
        '202':
          $ref: '#/components/responses/JobAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/MediaNotFound'
  /video/{mediaId}/detect-scenes:
    post:
      tags: [Video]
      summary: Crea un job para detectar escenas y guardar un artefacto scene_detection.
      operationId: detectScenes
      parameters:
        - $ref: '#/components/parameters/MediaId'
      requestBody:
        $ref: '#/components/requestBodies/VideoDetectScenes'
      responses:
        '202':
          $ref: '#/components/responses/JobAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/MediaNotFound'
  /video/{mediaId}/generate-thumbnails:
    post:
      tags: [Video]
      summary: Crea un job para generar thumbnails.
      operationId: generateThumbnails
      parameters:
        - $ref: '#/components/parameters/MediaId'
      requestBody:
        $ref: '#/components/requestBodies/VideoGenerateThumbnails'
      responses:
        '202':
          $ref: '#/components/responses/JobAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/MediaNotFound'
...
components:
  parameters:
    MediaId:
      name: mediaId
      in: path
      required: true
      schema:
        type: string
      example: med_abc123
...
  requestBodies:
...
    VideoExtractFrames:
      required: false
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VideoExtractFramesOptions'
    VideoDetectScenes:
      required: false
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VideoDetectScenesOptions'
    VideoGenerateThumbnails:
      required: false
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VideoGenerateThumbnailsOptions'
  responses:
    JobAccepted:
      description: Job creado correctamente.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JobResponse'
    BadRequest:
      description: Solicitud inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    MediaNotFound:
      description: Archivo no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            message: Archivo no encontrado.
            errorCode: MEDIA_NOT_FOUND
...
  schemas:
    ApiResponse:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
        message:
          type: string
        data:
          type: object
          additionalProperties: true
          nullable: true
    ErrorResponse:
      type: object
      required: [ok, message, errorCode]
      properties:
        ok:
          type: boolean
          example: false
        message:
          type: string
        errorCode:
          type: string
          enum:
            - INVALID_FILE
            - FILE_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
            - INVALID_MEDIA_FORMAT
            - MEDIA_NOT_FOUND
            - MEDIA_HAS_ARTIFACTS
            - ARTIFACT_NOT_FOUND
            - ARTIFACT_HAS_CHILDREN
            - JOB_NOT_FOUND
            - PROCESSING_FAILED
            - FFMPEG_FAILED
            - FFPROBE_FAILED
            - SIGNED_URL_FAILED
            - VALIDATION_FAILED
            - RATE_LIMITED
            - STORAGE_WRITE_FAILED
            - STORAGE_READ_FAILED
            - ROUTE_NOT_FOUND
...
    JobType:
      type: string
      enum:
        - media.extractMetadata
        - audio.transcode
        - audio.chunk
        - audio.normalize
        - audio.removeSilence
        - video.extractAudio
        - video.transcode
        - video.compress
        - video.split
        - video.extractFrames
        - video.detectScenes
        - video.generateThumbnails
    JobStatus:
      type: string
      enum: [queued, active, completed, failed, cancelled]
...
    Job:
      type: object
      required: [jobId, mediaId, type, status, progress]
      properties:
        jobId:
          type: string
          example: job_t4BrW8EnC1DKx5Ez
        queueJobId:
          type: string
        mediaId:
          type: string
        sourceType:
          type: string
          enum: [media, artifact]
          default: media
        sourceId:
          type: string
        type:
          $ref: '#/components/schemas/JobType'
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        currentStep:
          type: string
          example: processing_ffmpeg
        options:
          type: object
          additionalProperties: true
        result:
          type: object
          additionalProperties: true
          example:
            artifactCount: 3
            artifactType: thumbnail
        error:
          type: object
          nullable: true
          properties:
            message:
              type: string
            code:
              type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        failedAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    JobResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Job'
...
    VideoExtractFramesOptions:
      type: object
      properties:
        mode:
          type: string
          enum: [interval, scene-change, keyframes, hybrid]
          default: scene-change
        sceneThreshold:
          type: number
          minimum: 0
          maximum: 1
          default: 0.3
        intervalSeconds:
          type: integer
          minimum: 1
          maximum: 3600
          default: 30
        maxFrames:
          type: integer
          minimum: 1
          maximum: 500
          default: 30
        format:
          type: string
          enum: [png, jpeg, jpg, webp]
          default: png
        width:
          type: integer
          minimum: 64
          maximum: 7680
    VideoDetectScenesOptions:
      type: object
      properties:
        sceneThreshold:
          type: number
          minimum: 0
          maximum: 1
          default: 0.3
        minSceneDurationSeconds:
          type: number
          minimum: 0
          maximum: 60
          default: 2
    VideoGenerateThumbnailsOptions:
      type: object
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 50
          default: 5
        width:
          type: integer
          minimum: 64
          maximum: 1920
          default: 320
        format:
          type: string
          enum: [webp, png, jpeg, jpg]
          default: webp